refactor: 커스텀 체크리스트 수정 메서드 이전
