+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/controller/ChecklistController.java
@@ -84,7 +84,7 @@ public class ChecklistController {
     @PutMapping("/custom-checklist")
     public ResponseEntity<Void> updateCustomChecklist(@AuthPrincipal User user,
                                                       @RequestBody CustomChecklistUpdateRequest request) {
-        checklistService.updateCustomChecklist(user, request);
+        checklistManageService.updateCustomChecklist(user, request);
         return ResponseEntity.noContent().build();
     }
